# rev=true in searchsorted_interval