this also simplifies findall

no new tests added: rely on the extensive findall testsuite