@hyrodium 2022-09-23
Maybe, it seems better to split this PR into

* Add `rev` keyword argument in `searchsorted_interval` function.
* Simplify `findall` function.

to merge this PR quickly.